Adding unit tests for spark splitters
